refactor: 현장톡 말풍선 UI 분리
