+++ android/app/src/main/java/com/yagubogu/presentation/livetalk/chat/LivetalkMyBubbleViewHolder.kt
@@ -1,0 +1,22 @@
+package com.yagubogu.presentation.livetalk.chat
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.yagubogu.databinding.ItemLivetalkMyBubbleBinding
+
+class LivetalkMyBubbleViewHolder(
+    private val binding: ItemLivetalkMyBubbleBinding,
+) : RecyclerView.ViewHolder(binding.root) {
+    fun bind(item: LivetalkChatItem) {
+        binding.livetalkChatItem = item
+    }
+
+    companion object {
+        fun from(parent: ViewGroup): LivetalkMyBubbleViewHolder {
+            val inflater = LayoutInflater.from(parent.context)
+            val binding = ItemLivetalkMyBubbleBinding.inflate(inflater, parent, false)
+            return LivetalkMyBubbleViewHolder(binding)
+        }
+    }
+}
