+++ android/app/src/main/java/com/yagubogu/presentation/livetalk/chat/LivetalkOtherBubbleViewHolder.kt
@@ -1,0 +1,22 @@
+package com.yagubogu.presentation.livetalk.chat
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import com.yagubogu.databinding.ItemLivetalkOtherBubbleBinding
+
+class LivetalkOtherBubbleViewHolder(
+    private val binding: ItemLivetalkOtherBubbleBinding,
+) : RecyclerView.ViewHolder(binding.root) {
+    fun bind(item: LivetalkChatItem) {
+        binding.livetalkChatItem = item
+    }
+
+    companion object {
+        fun from(parent: ViewGroup): LivetalkOtherBubbleViewHolder {
+            val inflater = LayoutInflater.from(parent.context)
+            val binding = ItemLivetalkOtherBubbleBinding.inflate(inflater, parent, false)
+            return LivetalkOtherBubbleViewHolder(binding)
+        }
+    }
+}
